ledger-api-test-tool: Extract non-core test functionality into objects.
