+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/Divulgence.scala
@@ -4,8 +4,8 @@
 package com.daml.ledger.api.testtool.tests
 
 import com.daml.ledger.api.testtool.infrastructure.{LedgerSession, LedgerTest, LedgerTestSuite}
-import com.digitalasset.ledger.test_stable.Test.{Divulgence1, Divulgence2}
 import com.digitalasset.ledger.test_stable.Test.Divulgence2._
+import com.digitalasset.ledger.test_stable.Test.{Divulgence1, Divulgence2}
 import scalaz.Tag
 
 final class Divulgence(session: LedgerSession) extends LedgerTestSuite(session) {
