+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/PackageManagement.scala
@@ -3,6 +3,7 @@
 
 package com.daml.ledger.api.testtool.tests
 
+import com.daml.ledger.api.testtool.infrastructure.Assertions._
 import com.daml.ledger.api.testtool.infrastructure.{LedgerSession, LedgerTest, LedgerTestSuite}
 import com.digitalasset.ledger.packagemanagementtest.PackageManagementTest.PackageManagementTestTemplate
 import com.digitalasset.ledger.packagemanagementtest.PackageManagementTest.PackageManagementTestTemplate._
